+++ pandas/tests/frame/methods/test_to_csv.py
@@ -62,8 +62,8 @@ class TestDataFrameToCSV:
             # corner case
             dm = DataFrame(
                 {
-                    "s1": Series(range(3), index=np.arange(3)),
-                    "s2": Series(range(2), index=np.arange(2)),
+                    "s1": Series(range(3), index=np.arange(3, dtype=np.int64)),
+                    "s2": Series(range(2), index=np.arange(2, dtype=np.int64)),
                 }
             )
             dm.to_csv(path)
@@ -388,7 +388,7 @@ class TestDataFrameToCSV:
 
     @pytest.mark.slow
     def test_to_csv_empty(self):
-        df = DataFrame(index=np.arange(10))
+        df = DataFrame(index=np.arange(10, dtype=np.int64))
         result, expected = self._return_result_expected(df, 1000)
         tm.assert_frame_equal(result, expected, check_column_type=False)
 
@@ -486,7 +486,7 @@ class TestDataFrameToCSV:
 
         frame = float_frame
         old_index = frame.index
-        arrays = np.arange(len(old_index) * 2).reshape(2, -1)
+        arrays = np.arange(len(old_index) * 2, dtype=np.int64).reshape(2, -1)
         new_index = MultiIndex.from_arrays(arrays, names=["first", "second"])
         frame.index = new_index
 
@@ -510,7 +510,7 @@ class TestDataFrameToCSV:
             # try multiindex with dates
             tsframe = datetime_frame
             old_index = tsframe.index
-            new_index = [old_index, np.arange(len(old_index))]
+            new_index = [old_index, np.arange(len(old_index), dtype=np.int64)]
             tsframe.index = MultiIndex.from_arrays(new_index)
 
             tsframe.to_csv(path, index_label=["time", "foo"])
